jbeuhler-flexelint-fun-with-printf-20031128
This patch addresses a number of problems with printf-like functions.

Most of them are %x formats applied to pointers (which could be a
problem depending on the architecture I would assume).  There is a
%p format for pointers on many machines but I don't know if it
applies to all supported OpenAFS platforms so I didn't go that route.

There is a buffer overflow fix to ktime.c

There are a couple places where incorrect format characters were used.

One call had two arguments swapped.

Other fixes include the use of PrintInode() and %s formats for
printing inode numbers.  The code was inconsistent in this regard -- sometimes
PrintInode() was used, sometimes it was side-stepped.

There is one place where a local shadowed another variable, so I renamed
the local.

The last segment points out that an fdset need not be anything printable
with an integer format.
--
Joe Buehler

Comments - The fdset comment relates to a WinSock only file on which
the size is a pointer to int.  Modified the source to use a %p instead
of %x as it is supported on Windows.

Added an assert(p != NULL) to src/rx/rx_packet.c

+++ src/WINNT/afsd/afsd_init.c
@@ -464,7 +464,7 @@ int afsd_InitCM(char **reasonP)
         
 	/* setup and enable debug log */
 	afsd_logp = osi_LogCreate("afsd", traceBufSize);
-	afsi_log("osi_LogCreate log addr %x", afsd_logp);
+	afsi_log("osi_LogCreate log addr %x", (int)afsd_logp);
         osi_LogEnable(afsd_logp);
 	logReady = 1;
 
@@ -510,7 +510,7 @@ int afsd_InitCM(char **reasonP)
 	nullServerSecurityClassp = rxnull_NewServerSecurityObject();
         serverp = rx_NewService(0, 1, "AFS", &nullServerSecurityClassp, 1,
         	RXAFSCB_ExecuteRequest);
-	afsi_log("rx_NewService addr %x", serverp);
+	afsi_log("rx_NewService addr %x", (int)serverp);
 	if (serverp == NULL) {
 		*reasonP = "unknown error";
 		return -1;
@@ -519,7 +519,7 @@ int afsd_InitCM(char **reasonP)
 	nullServerSecurityClassp = rxnull_NewServerSecurityObject();
         serverp = rx_NewService(0, RX_STATS_SERVICE_ID, "rpcstats",
 		&nullServerSecurityClassp, 1, RXSTATS_ExecuteRequest);
-	afsi_log("rx_NewService addr %x", serverp);
+	afsi_log("rx_NewService addr %x", (int)serverp);
 	if (serverp == NULL) {
 		*reasonP = "unknown error";
 		return -1;
@@ -575,7 +575,7 @@ int afsd_InitCM(char **reasonP)
 
         if (code == 0 && !cm_freelanceEnabled) {
 	  cm_rootCellp = cm_GetCell(rootCellName, CM_FLAG_CREATE);
-          afsi_log("cm_GetCell addr %x", cm_rootCellp);
+          afsi_log("cm_GetCell addr %x", (int)cm_rootCellp);
 	  if (cm_rootCellp == NULL) {
 	    *reasonP = "can't find root cell in afsdcell.ini";
 	    return -1;
